Таки разобрался с созданием базы один раз для всех тестов

+++ tests/test_statistic.py
@@ -11,6 +11,8 @@ import database as db
 import statistic
 import datetime as dtime
 
+first_run: bool = True
+
 class CTestStatistic(TestCase):
 
     def setUp(self) -> None:
@@ -24,20 +26,19 @@ class CTestStatistic(TestCase):
         else:
 
             self.data_path: str = self.config[softice.WINDOWS_DATA_FOLDER_KEY]
-        """
-        Как-то нужно удалять и создавать базу один раз, а потом уже не дрыгаться
-        for file in Path(self.data_path).glob("softice.db"):
-
-           file.unlink()
-
-        print(f"{self.data_path=}")    
-        file_name =  Path(self.data_path) / "softice.db"
-        if not file_name.is_file():
-
-            print("Create!")
-            self.database.create()
-        """            
         self.database: db.CDataBase = db.CDataBase(self.config, self.data_path)
+        global first_run    
+        if first_run:
+
+            for file in Path(self.data_path).glob("softice.db"):
+
+               file.unlink()
+            file_name =  Path(self.data_path) / "softice.db"
+            if not file_name.is_file():
+
+                print("Create!")
+                self.database.create()
+            first_run = False    
         self.statistic: statistic.CStatistic = statistic.CStatistic(self.config, self.database)
